+++ lib/src/course/course_component.dart
@@ -1,0 +1,7 @@
+library plato.archives.components.course;
+
+/// The [CourseComponent] class...
+class CourseComponent {
+  /// The [CourseComponent] constructor...
+  CourseComponent();
+}