+++ lib/src/enrollment/enrollment.dart
@@ -1,0 +1,13 @@
+library plato.archives.models.enrollment;
+
+/// The [Enrollment] class...
+class Enrollment {
+  final String id;
+
+  final String courseId;
+
+  final String userId;
+
+  /// The [Enrollment] constructor...
+  Enrollment (this.id, this.courseId, this.userId);
+}