+++ lib/src/enrollment/roster.dart
@@ -1,0 +1,7 @@
+library plato.archives.models.enrollment.roster;
+
+/// The [Roster] class...
+class Roster {
+  /// The [Roster] constructor...
+  Roster();
+}