+++ lib/src/plato_archives.dart
@@ -2,11 +2,13 @@ library plato.archives.components.plato_archives;
 
 import 'package:angular/angular.dart';
 
+/// The [PlatoArchives] class...
 @Component(
   selector: 'plato-archives',
   templateUrl: 'plato_archives.html',
   directives: [],
 )
 class PlatoArchives {
+  /// The [PlatoArchives] constructor...
   PlatoArchives();
 }