+++ lib/src/resource/resource.dart
@@ -1,0 +1,13 @@
+library plato.archives.models.resource;
+
+/// The [Resource] class...
+class Resource {
+  final String id;
+
+  final String title;
+
+  final String type;
+
+  /// The [Resource] constructor...
+  Resource (this.id, this.title, this.type);
+}