+++ lib/src/resource/resource_component.dart
@@ -1,0 +1,7 @@
+library plato.archives.components.resource;
+
+/// The [ResourceComponent] class...
+class ResourceComponent {
+  /// The [ResourceComponent] constructor...
+  ResourceComponent();
+}