+++ lib/src/user/authentication_component.dart
@@ -1,0 +1,11 @@
+library plato.archives.components.user.authentication;
+
+import 'authentication_service.dart';
+
+/// The [AuthenticationComponent] class...
+class AuthenticationComponent {
+  final AuthenticationService _authService;
+
+  /// The [AuthenticationComponent] constructor...
+  AuthenticationComponent (this._authService);
+}